% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mht-package.R
\docType{package}
\name{mht-package}
\alias{mht}
\alias{mht-package}
\title{mht: Menopausal hormone therapy exposure definitions for Swedish registry data}
\description{
Derives menopausal hormone therapy (MHT) exposure from Swedish prescription
registry (LMED) data. Categorises dispensed product names into MHT groups,
corrects dispensed durations, bridges gaps between consecutive prescriptions,
and applies the approach-based exposure definitions used by the
Skalkidou-lab MHT studies to person-week skeletons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/skalkidou-lab/mht}
  \item \url{https://skalkidou-lab.github.io/mht/}
  \item Report bugs at \url{https://github.com/skalkidou-lab/mht/issues}
}

}
\author{
\strong{Maintainer}: Richard Aubrey White \email{hello@rwhite.no} (\href{https://orcid.org/0000-0002-6747-1726}{ORCID})

Authors:
\itemize{
  \item Richard Aubrey White \email{hello@rwhite.no} (\href{https://orcid.org/0000-0002-6747-1726}{ORCID})
}

}
\keyword{internal}
